split things into feature folders
