+++ src/slskd/API/Controllers/ConversationsController.cs
@@ -24,8 +24,7 @@ namespace slskd.API.Controllers
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
     using slskd.API.DTO;
-    using slskd.Entities;
-    using slskd.Trackers;
+    using slskd.Messaging;
     using Soulseek;
 
     /// <summary>
