+++ src/slskd/API/Controllers/SearchesController.cs
@@ -25,7 +25,7 @@ namespace slskd.API.Controllers
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
     using slskd.API.DTO;
-    using slskd.Trackers;
+    using slskd.Search;
     using Soulseek;
 
     /// <summary>
