+++ src/slskd/API/Controllers/UserController.cs
@@ -23,7 +23,7 @@ namespace slskd.API.Controllers
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
     using slskd.API.DTO;
-    using slskd.Trackers;
+    using slskd.Peer;
     using Soulseek;
 
     /// <summary>
