+++ src/slskd/Local/SharedFileCache.cs
@@ -15,7 +15,7 @@
 //     along with this program.  If not, see https://www.gnu.org/licenses/.
 // </copyright>
 
-namespace slskd
+namespace slskd.Local
 {
     using System;
     using System.Collections.Generic;
