+++ src/slskd/Messaging/Entities/Room.cs
@@ -15,7 +15,7 @@
 //     along with this program.  If not, see https://www.gnu.org/licenses/.
 // </copyright>
 
-namespace slskd.Entities
+namespace slskd.Messaging
 {
     using System.Collections.Generic;
     using System.Linq;
