+++ src/slskd/Search/SearchResponseCache.cs
@@ -15,7 +15,7 @@
 //     along with this program.  If not, see https://www.gnu.org/licenses/.
 // </copyright>
 
-namespace slskd
+namespace slskd.Search
 {
     using System;
     using Microsoft.Extensions.Caching.Memory;
